KAFKA-8943: Move SecurityProviderCreator to org.apache.kafka.common.security.auth package (#7564)
Reviewers: Mickael Maison <mickael.maison@gmail.com>, Rajini Sivaram <rajinisivaram@googlemail.com>
