+++ clients/src/main/java/org/apache/kafka/common/config/SecurityConfig.java
@@ -22,7 +22,8 @@ package org.apache.kafka.common.config;
 public class SecurityConfig {
 
     public static final String SECURITY_PROVIDERS_CONFIG = "security.providers";
-    public static final String SECURITY_PROVIDERS_DOC = "A list of configurable creators each returning a provider " +
-            "implementing security algorithms";
+    public static final String SECURITY_PROVIDERS_DOC = "A list of configurable creator classes each returning a provider" +
+        " implementing security algorithms. These classes should implement the" +
+        " <code>org.apache.kafka.common.security.auth.SecurityProviderCreator</code> interface.";
 
 }
