+++ clients/src/main/java/org/apache/kafka/common/security/auth/SecurityProviderCreator.java
@@ -14,9 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.security;
+package org.apache.kafka.common.security.auth;
 
 import org.apache.kafka.common.Configurable;
+import org.apache.kafka.common.annotation.InterfaceStability;
 
 import java.security.Provider;
 import java.util.Map;
@@ -24,6 +25,7 @@ import java.util.Map;
 /**
  * An interface for generating security providers.
  */
+@InterfaceStability.Evolving
 public interface SecurityProviderCreator extends Configurable {
 
     /**
